backend/requirements.txt
Flask==3.1.3
flask-cors==6.0.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
pymongo==4.17.0
python-dotenv==1.2.2
openai==2.33.0
gunicorn==25.3.0